+++ braindecode/functional/functions.py
@@ -1,7 +1,6 @@
 # Authors: Robin Schirrmeister <robintibor@gmail.com>
 #
 # License: BSD (3-clause)
-import math
 
 import torch
 import torch.nn.functional as F