+++ braindecode/models/__init__.py
@@ -4,21 +4,20 @@ Some predefined network architectures for EEG decoding.
 
 from .atcnet import ATCNet
 from .attentionbasenet import AttentionBaseNet
+from .attn_sleep import AttnSleep
 from .base import EEGModuleMixin
 from .biot import BIOT
 from .contrawr import ContraWR
 from .ctnet import CTNet
 from .deep4 import Deep4Net
-from .deeprecurrent import DeepRecurrent
 from .deepsleepnet import DeepSleepNet
 from .eegconformer import EEGConformer
 from .eeginception_erp import EEGInceptionERP
 from .eeginception_mi import EEGInceptionMI
 from .eegitnet import EEGITNet
 from .eegminer import EEGMiner
-from .eegnet import EEGNetv1, EEGNetv4
+from .eegnet import EEGNet, EEGNetv4
 from .eegnex import EEGNeX
-from .eegresnet import EEGResNet
 from .eegsimpleconv import EEGSimpleConv
 from .eegtcnet import EEGTCNet
 from .fbcnet import FBCNet
@@ -28,6 +27,7 @@ from .hybrid import HybridNet
 from .ifnet import IFNet
 from .labram import Labram
 from .msvtnet import MSVTNet
+from .patchedtransformer import PBT
 from .sccnet import SCCNet
 from .shallow_fbcsp import ShallowFBCSPNet
 from .signal_jepa import (
@@ -39,15 +39,64 @@ from .signal_jepa import (
 from .sinc_shallow import SincShallowNet
 from .sleep_stager_blanco_2020 import SleepStagerBlanco2020
 from .sleep_stager_chambon_2018 import SleepStagerChambon2018
-from .sleep_stager_eldele_2021 import SleepStagerEldele2021
 from .sparcnet import SPARCNet
+from .sstdpn import SSTDPN
 from .syncnet import SyncNet
 from .tcn import BDTCN, TCN
 from .tidnet import TIDNet
-from .tsinception import TSceptionV1
+from .tsinception import TSception
 from .usleep import USleep
 from .util import _init_models_dict, models_mandatory_parameters
 
 # Call this last in order to make sure the dataset list is populated with
 # the models imported in this file.
 _init_models_dict()
+
+__all__ = [
+    "ATCNet",
+    "AttnSleep",
+    "AttentionBaseNet",
+    "EEGModuleMixin",
+    "BIOT",
+    "ContraWR",
+    "CTNet",
+    "Deep4Net",
+    "DeepSleepNet",
+    "EEGConformer",
+    "EEGInceptionERP",
+    "EEGInceptionMI",
+    "EEGITNet",
+    "EEGMiner",
+    "EEGNet",
+    "EEGNetv4",
+    "EEGNeX",
+    "EEGSimpleConv",
+    "EEGTCNet",
+    "FBCNet",
+    "FBLightConvNet",
+    "FBMSNet",
+    "HybridNet",
+    "IFNet",
+    "Labram",
+    "MSVTNet",
+    "PBT",
+    "SCCNet",
+    "ShallowFBCSPNet",
+    "SignalJEPA",
+    "SignalJEPA_Contextual",
+    "SignalJEPA_PostLocal",
+    "SignalJEPA_PreLocal",
+    "SincShallowNet",
+    "SSTDPN",
+    "SleepStagerBlanco2020",
+    "SleepStagerChambon2018",
+    "SPARCNet",
+    "SyncNet",
+    "BDTCN",
+    "TCN",
+    "TIDNet",
+    "TSception",
+    "USleep",
+    "_init_models_dict",
+    "models_mandatory_parameters",
+]