+++ braindecode/models/eegminer.py
@@ -10,7 +10,6 @@ from functools import partial
 import torch
 from einops.layers.torch import Rearrange
 from torch import nn
-from torch.fft import fftfreq
 
 import braindecode.functional as F
 from braindecode.models.base import EEGModuleMixin