+++ test/unit_tests/models/test_models.py
@@ -21,6 +21,7 @@ from braindecode.models import (
     TCN,
     ATCNet,
     AttentionBaseNet,
+    AttnSleep,
     ContraWR,
     Deep4Net,
     DeepSleepNet,
@@ -29,9 +30,7 @@ from braindecode.models import (
     EEGInceptionMI,
     EEGITNet,
     EEGMiner,
-    EEGNetv1,
-    EEGNetv4,
-    EEGResNet,
+    EEGNet,
     EEGNeX,
     EEGSimpleConv,
     EEGTCNet,
@@ -39,12 +38,13 @@ from braindecode.models import (
     HybridNet,
     IFNet,
     Labram,
+    SCCNet,
     ShallowFBCSPNet,
     SleepStagerBlanco2020,
     SleepStagerChambon2018,
-    SleepStagerEldele2021,
     SPARCNet,
     TIDNet,
+    TSception,
     USleep,
 )
 from braindecode.util import set_random_seeds
@@ -86,6 +86,25 @@ def check_forward_pass(model, input_sizes, only_check_until_dim=None):
         y_pred_new.detach().cpu().numpy(),
         atol=1e-4,
         rtol=0,
+    )
+
+def check_forward_pass_3d(model, input_sizes, only_check_until_dim=None):
+    rng = np.random.RandomState(42)
+
+    # Test 3d input
+    X = rng.randn(
+        input_sizes["n_samples"],
+        input_sizes["n_channels"],
+        input_sizes["n_in_times"],
+    )
+    X = torch.Tensor(X.astype(np.float32))
+    set_random_seeds(0, False)
+    X = X.squeeze(-1)
+    assert len(X.shape) == 3
+    y_pred_new = model(X)
+    assert y_pred_new.shape[:only_check_until_dim] == (
+        input_sizes["n_samples"],
+        input_sizes["n_classes"],
     )
 
 
@@ -176,27 +195,6 @@ def test_deep4net_load_state_dict(input_sizes):
     model.load_state_dict(state_dict)
 
 
-def test_eegresnet(input_sizes):
-    model = EEGResNet(
-        input_sizes["n_channels"],
-        input_sizes["n_classes"],
-        input_sizes["n_in_times"],
-        final_pool_length=5,
-        n_first_filters=2,
-    )
-    check_forward_pass(model, input_sizes, only_check_until_dim=2)
-
-
-def test_eegresnet_pool_length_auto(input_sizes):
-    model = EEGResNet(
-        input_sizes["n_channels"],
-        input_sizes["n_classes"],
-        input_sizes["n_in_times"],
-        final_pool_length="auto",
-        n_first_filters=2,
-    )
-    check_forward_pass(model, input_sizes, only_check_until_dim=2)
-
 
 def test_hybridnet(input_sizes):
     model = HybridNet(
@@ -207,25 +205,13 @@ def test_hybridnet(input_sizes):
     check_forward_pass(model, input_sizes, only_check_until_dim=2)
 
 
-def test_eegnet_v4(input_sizes):
-    model = EEGNetv4(
+def test_eegnet(input_sizes):
+    model = EEGNet(
         input_sizes["n_channels"],
         input_sizes["n_classes"],
         n_times=input_sizes["n_in_times"],
     )
     check_forward_pass(model, input_sizes)
-
-
-def test_eegnet_v1(input_sizes):
-    model = EEGNetv1(
-        n_chans=input_sizes["n_channels"],
-        n_outputs=input_sizes["n_classes"],
-        n_times=input_sizes["n_in_times"],
-    )
-    check_forward_pass(
-        model,
-        input_sizes,
-    )
 
 
 def test_tcn(input_sizes):
@@ -510,7 +496,7 @@ def test_eldele_2021(sfreq, n_classes, input_size_s, d_model):
     n_channels = 1
     n_examples = 10
 
-    model = SleepStagerEldele2021(
+    model = AttnSleep(
         sfreq=sfreq,
         n_outputs=n_classes,
         input_window_seconds=input_size_s,
@@ -534,7 +520,7 @@ def test_eldele_2021_feats():
     n_classes = 3
     n_examples = 10
 
-    model = SleepStagerEldele2021(
+    model = AttnSleep(
         sfreq,
         input_window_seconds=input_size_s,
         n_outputs=n_classes,
@@ -720,6 +706,241 @@ def test_model_creation(model):
     assert model is not None
 
 
+def test_sparcnet_dummy():
+    input_sizes = dict(n_channels=32, n_in_times=125, n_classes=2, n_samples=64)
+    model = SPARCNet(
+        n_chans=input_sizes["n_channels"],
+        n_outputs=input_sizes["n_classes"],
+        n_times=input_sizes["n_in_times"],
+        sfreq=500.0,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (256, 8, 256.0, 2),
+        (204, 8, 256.0, 2),
+        (125, 32, 500.0, 2),
+        (204, 16, 256.0, 2),
+        (128, 16, 128.0, 2),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_atcnet_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = ATCNet(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (125, 32, 500.0, 2),
+        (614, 64, 2048.0, 2),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_tsception_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = TSception(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (125, 32, 500.0, 2),
+        (614, 64, 2048.0, 2),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_sccnet_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = SCCNet(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (2000, 63, 500.0, 4),
+    ],
+)
+def test_eeginceptionmi_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = EEGInceptionMI(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        input_window_seconds=n_times / sfreq,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (256, 8, 256.0, 2),
+        (204, 8, 256.0, 2),
+        (125, 32, 500.0, 2),
+        (204, 16, 256.0, 2),
+        (128, 16, 128.0, 2),
+        (384, 14, 128.0, 5),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_deep4net_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = Deep4Net(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (1536, 16, 512.0, 3),
+        (512, 16, 512.0, 2),
+        (125, 32, 500.0, 2),
+        (2560, 32, 512.0, 2),
+        (2560, 13, 512.0, 2),
+        (512, 32, 512.0, 2),
+        (2560, 15, 512.0, 2),
+        (1024, 30, 1024.0, 2),
+        (5120, 16, 512.0, 2),
+        (1536, 64, 512.0, 2),
+        (2048, 32, 2048.0, 2),
+        (614, 64, 2048.0, 2),
+        (1536, 61, 512.0, 7),
+        (899, 31, 1000.0, 2),
+        (4000, 62, 1000.0, 4),
+        (4000, 62, 1000.0, 2),
+        (153, 8, 512.0, 2),
+        (1000, 62, 1000.0, 2),
+        (1200, 31, 1000.0, 2),
+    ],
+)
+def test_contrawr_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = ContraWR(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (204, 8, 256.0, 2),
+        (125, 32, 500.0, 2),
+        (204, 16, 256.0, 2),
+        (614, 64, 2048.0, 2),
+        (899, 31, 1000.0, 2),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_biot_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = BIOT(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
+@pytest.mark.parametrize(
+    "n_times, n_chans, sfreq, n_outputs",
+    [
+        (125, 32, 500.0, 2),
+        (128, 16, 128.0, 2),
+        (153, 8, 512.0, 2),
+    ],
+)
+def test_attentionbasenet_dummy(n_times, n_chans, sfreq, n_outputs):
+    batch_size = 64
+    input_sizes = dict(
+        n_channels=n_chans,
+        n_in_times=n_times,
+        n_classes=n_outputs,
+        n_samples=batch_size,
+    )
+    model = AttentionBaseNet(
+        n_chans=n_chans,
+        n_outputs=n_outputs,
+        n_times=n_times,
+        sfreq=sfreq,
+    )
+    check_forward_pass_3d(model, input_sizes)
+
+
 
 def test_conformer_forward_pass(sample_input, model):
     output = model(sample_input)
@@ -803,6 +1024,7 @@ def default_biot_params():
         "hop_length": 50,
         "n_outputs": 2,
         "n_chans": 64,
+        "n_times": 1000,
     }
 
 
@@ -1072,7 +1294,7 @@ def test_attentionbasenet(default_attentionbasenet_params, attention_mode):
 
 def test_parameters_contrawr():
 
-    model = ContraWR(n_outputs=2, n_chans=22, sfreq=250)
+    model = ContraWR(n_outputs=2, n_chans=22, sfreq=250, n_times=1000)
 
     n_params = sum(p.numel() for p in model.parameters() if p.requires_grad)
     # 1.6M parameters according to the Labram paper, table 1
@@ -1248,9 +1470,9 @@ def test_eegminer_plv_values_range():
         "PLV values should be in the range [0, 1]"
 
 
-def test_eegnetv4_final_layer_linear_true():
+def test_eegnet_final_layer_linear_true():
     """Test that final_layer_linear=True uses a conv-based classifier without warning."""
-    model = EEGNetv4(
+    model = EEGNet(
         final_layer_with_constraint=True,
         n_chans=4,
         n_times=128,
@@ -1269,12 +1491,12 @@ def test_eegnetv4_final_layer_linear_true():
     assert hasattr(final_layer,
                    "linearconstraint"), "Expected a 'linear constraint' sub-module."
 
-def test_eegnetv4_final_layer_linear_false():
+def test_eegnet_final_layer_linear_false():
     """Test that final_layer_conv=False raises a DeprecationWarning and uses
     a linear layer."""
     with pytest.warns(DeprecationWarning,
                       match="Parameter 'final_layer_with_constraint=False' is deprecated"):
-        model = EEGNetv4(
+        model = EEGNet(
             final_layer_with_constraint=False,
             n_chans=4,
             n_times=128,
@@ -1474,3 +1696,29 @@ def test_eegnex_final_layer_in_features(n_times_input):
     )
 
     print(model)
+
+@pytest.mark.parametrize("batch_norm", [True, False])
+def test_batchnorm_deep4net(batch_norm):
+    """
+    Test the number of trainable parameters in Deep4Net model.
+    """
+    model = Deep4Net(n_outputs=2, n_chans=22, n_times=1000, batch_norm=batch_norm)
+
+    assert model is not None
+
+def test_fc_length_eegconformer():
+    """
+    Test the number of trainable parameters in EEGConformer model.
+    """
+    model = EEGConformer(
+        n_chans=64,  # Number of EEG channels
+        n_outputs=2,  # Number of output classes
+        n_times=500,  # Length of the input sequence (e.g., 500 time steps)
+        final_fc_length=120,
+        input_window_seconds=1.0,
+        return_features=True,
+        drop_prob=0.5,  # Dropout probability
+        sfreq=500.0  # Sampling frequency of the EEG data
+    )
+
+    assert model is not None